qc/utils: relocate 'determine_qc_protocol' to the 'protocols' module.
