+++ auto_process_ngs/qc/protocols.py
@@ -41,6 +41,7 @@ The available modifiers are the same as the parameter list for the
 
 This module also provides the following functions:
 
+- determine_qc_protocol: get QC protocol for a project
 - fetch_protocol_definition: get the definition for a QC protocol
 """
 
@@ -195,11 +196,79 @@ QC_PROTOCOLS = {
             'strandedness'
         ]
     },
+
+    "ICELL8_scATAC": {
+        "reads": {
+            "data": ('r1','r2',),
+            "qc": ('r1','r2',)
+        },
+        "qc_modules": [
+            'fastqc',
+            'fastq_screen',
+            'sequence_lengths',
+            'strandedness'
+        ]
+    },
 }
 
 #######################################################################
 # Functions
 #######################################################################
+
+def determine_qc_protocol(project):
+    """
+    Determine the QC protocol for a project
+
+    Arguments:
+      project (AnalysisProject): project instance
+
+    Return:
+      String: QC protocol for the project
+    """
+    # Standard protocols
+    if project.info.paired_end:
+        protocol = "standardPE"
+    else:
+        protocol = "standardSE"
+    # Single cell protocols
+    if project.info.single_cell_platform is not None:
+        # Default
+        protocol = "singlecell"
+        single_cell_platform = project.info.single_cell_platform
+        library_type = project.info.library_type
+        if single_cell_platform.startswith('10xGenomics Chromium 3\''):
+            if library_type == "scRNA-seq":
+                # 10xGenomics scATAC-seq
+                protocol = "10x_scRNAseq"
+            elif library_type == "snRNA-seq":
+                # 10xGenomics snRNA-seq
+                protocol = "10x_snRNAseq"
+            elif library_type in ("CellPlex",
+                                  "CellPlex scRNA-seq",
+                                  "CellPlex snRNA-seq"):
+                # 10xGenomics CellPlex (cell multiplexing)
+                protocol = "10x_CellPlex"
+        elif library_type in ("scATAC-seq",
+                              "snATAC-seq",):
+            if single_cell_platform == "10xGenomics Single Cell ATAC":
+                # 10xGenomics scATAC-seq
+                protocol = "10x_scATAC"
+            elif single_cell_platform == "ICELL8":
+                # ICELL8 scATAC-seq
+                protocol = "ICELL8_scATAC"
+    # Spatial RNA-seq
+    if project.info.single_cell_platform == "10xGenomics Visium":
+        # 10xGenomics Visium spatial transcriptomics
+        protocol = "10x_Visium"
+    # Multiome ATAC+GEX
+    if project.info.single_cell_platform == "10xGenomics Single Cell Multiome":
+        if library_type == "ATAC":
+            # 10xGenomics single cell Multiome ATAC
+            protocol = "10x_Multiome_ATAC"
+        elif library_type == "GEX":
+            # 10xGenomics single cell Multiome gene expression
+            protocol = "10x_Multiome_GEX"
+    return protocol
 
 def fetch_protocol_definition(name):
     """
