+++ auto_process_ngs/test/qc/test_utils.py
@@ -6,7 +6,6 @@ import unittest
 import tempfile
 import shutil
 import os
-from bcftbx.JobRunner import SimpleJobRunner
 from bcftbx.utils import mkdirs
 from auto_process_ngs.analysis import AnalysisProject
 from auto_process_ngs.mock import MockAnalysisProject
@@ -18,390 +17,10 @@ from auto_process_ngs.mock10xdata import CELLPLEX_METRICS_SUMMARY
 from auto_process_ngs.mock10xdata import MULTIOME_SUMMARY
 from auto_process_ngs.qc.utils import verify_qc
 from auto_process_ngs.qc.utils import report_qc
-from auto_process_ngs.qc.utils import determine_qc_protocol
 from auto_process_ngs.qc.utils import set_cell_count_for_project
 
 # Set to False to keep test output dirs
 REMOVE_TEST_OUTPUTS = True
-
-class TestDetermineQCProtocolFunction(unittest.TestCase):
-    """
-    Tests for determine_qc_protocol function
-    """
-    def setUp(self):
-        # Create a temp working dir
-        self.wd = tempfile.mkdtemp(suffix='TestDetermineQCProtocolFunction')
-
-    def tearDown(self):
-        # Remove the temporary test directory
-        if REMOVE_TEST_OUTPUTS:
-            shutil.rmtree(self.wd)
-
-    def test_determine_qc_protocol_standardPE(self):
-        """determine_qc_protocol: standard paired-end run
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"))
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "standardPE")
-
-    def test_determine_qc_protocol_standardSE(self):
-        """determine_qc_protocol: standard single-end run
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",))
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "standardSE")
-
-    def test_determine_qc_protocol_icell8(self):
-        """determine_qc_protocol: single-cell run (ICELL8)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "ICELL8"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "singlecell")
-
-    def test_determine_qc_protocol_10xchromium3(self):
-        """determine_qc_protocol: single-cell run (10xGenomics Chromium 3')
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "singlecell")
-
-    def test_determine_qc_protocol_10xchromium3v2(self):
-        """determine_qc_protocol: single-cell run (10xGenomics Chromium 3'v2)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'v2"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "singlecell")
-
-    def test_determine_qc_protocol_10xchromium3v3(self):
-        """determine_qc_protocol: single-cell run (10xGenomics Chromium 3'v3)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'v3"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "singlecell")
-
-    def test_determine_qc_protocol_10xchromium3_rna_seq(self):
-        """determine_qc_protocol: single-cell RNA-seq (10xGenomics Chromium 3')
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'",
-                                          'Library type':
-                                          "scRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_scRNAseq")
-
-    def test_determine_qc_protocol_10xchromium3v2_rna_seq(self):
-        """determine_qc_protocol: single-cell RNA-seq (10xGenomics Chromium 3'v2)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'v2",
-                                          'Library type':
-                                          "scRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_scRNAseq")
-
-    def test_determine_qc_protocol_10xchromium3v3_rna_seq(self):
-        """determine_qc_protocol: single-cell RNA-seq (10xGenomics Chromium 3'v3)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'v3",
-                                          'Library type':
-                                          "scRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_scRNAseq")
-
-    def test_determine_qc_protocol_10xchromium3_sn_rna_seq(self):
-        """determine_qc_protocol: single-nuclei RNA-seq (10xGenomics Chromium 3')
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'",
-                                          'Library type':
-                                          "snRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_snRNAseq")
-
-    def test_determine_qc_protocol_10xchromium3v3_sn_rna_seq(self):
-        """determine_qc_protocol: single-nuclei RNA-seq (10xGenomics Chromium 3'v3)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'v3",
-                                          'Library type':
-                                          "snRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_snRNAseq")
-
-    def test_determine_qc_protocol_10xchromium3_cellplex(self):
-        """determine_qc_protocol: cell multiplexing CellPlex (10xGenomics Chromium 3')
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'",
-                                          'Library type':
-                                          "CellPlex"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_CellPlex")
-
-    def test_determine_qc_protocol_10xchromium3_cellplex_scrnaseq(self):
-        """determine_qc_protocol: cell multiplexing CellPlex scRNA-seq (10xGenomics Chromium 3')
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'",
-                                          'Library type':
-                                          "CellPlex scRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_CellPlex")
-
-    def test_determine_qc_protocol_10xchromium3_cellplex_snrnaseq(self):
-        """determine_qc_protocol: cell multiplexing CellPlex snRNA-seq (10xGenomics Chromium 3')
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'",
-                                          'Library type':
-                                          "CellPlex snRNA-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_CellPlex")
-
-    def test_determine_qc_protocol_10xchromium3v3_cellplex(self):
-        """determine_qc_protocol: cell multiplexing CellPlex (10xGenomics Chromium 3'v3)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Chromium 3'v3",
-                                          'Library type':
-                                          "CellPlex"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_CellPlex")
-
-    def test_determine_qc_protocol_10x_single_cell_atac_seq(self):
-        """determine_qc_protocol: single-cell ATAC-seq (10xGenomics Single Cell ATAC)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Single Cell ATAC",
-                                          'Library type':
-                                          "scATAC-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_scATAC")
-
-    def test_determine_qc_protocol_10x_single_nuclei_atac_seq(self):
-        """determine_qc_protocol: single-nuclei ATAC-seq (10xGenomics Single Cell ATAC)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Single Cell ATAC",
-                                          'Library type':
-                                          "snATAC-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_scATAC")
-
-    def test_determine_qc_protocol_icell8_atac_seq(self):
-        """determine_qc_protocol: single-cell ATAC-seq (ICELL8)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "ICELL8",
-                                          'Library type':
-                                          "scATAC-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "ICELL8_scATAC")
-
-    def test_determine_qc_protocol_10x_visium(self):
-        """determine_qc_protocol: spatial RNA-seq run (10xGenomics Visium)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Visium",
-                                          'Library type':
-                                          "scATAC-seq"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_Visium")
-
-    def test_determine_qc_protocol_10x_multiome_atac(self):
-        """determine_qc_protocol: single cell multiome ATAC run (10xGenomics Multiome ATAC)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"
-                                       "PJB2_S2_R3_001.fastq.gz",
-                                       "PJB2_S2_R3_001.fastq.gz",
-                                       "PJB1_S1_I1_001.fastq.gz",
-                                       "PJB1_S1_I2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Single Cell Multiome",
-                                          'Library type':
-                                          "ATAC"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_Multiome_ATAC")
-
-    def test_determine_qc_protocol_10x_multiome_gex(self):
-        """determine_qc_protocol: single cell multiome GEX run (10xGenomics Multiome GEX)
-        """
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz",
-                                       "PJB1_S1_I1_001.fastq.gz",
-                                       "PJB1_S1_I2_001.fastq.gz"),
-                                metadata={'Single cell platform':
-                                          "10xGenomics Single Cell Multiome",
-                                          'Library type':
-                                          "GEX"})
-        p.create(top_dir=self.wd)
-        project = AnalysisProject("PJB",
-                                  os.path.join(self.wd,"PJB"))
-        self.assertEqual(determine_qc_protocol(project),
-                         "10x_Multiome_GEX")
 
 class TestVerifyQCFunction(unittest.TestCase):
     """
